This fixes a bunch of the tests that were failing from me changing remove_file to only pass on OSError.
git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@7890 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: d64223b1a1ec2a379e5c0cb8de9027083b435424

+++ tv/portable/iconcache.py
@@ -140,7 +140,8 @@ class IconCache:
 
     def remove(self):
         self.removed = True
-        self.remove_file(self.filename)
+        if self.filename:
+            self.remove_file(self.filename)
 
     def reset(self):
         self.remove_file(self.filename)
